+++ lib/src/extensions/extensions.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
+export 'build_context_extension.dart';
 export 'date_time_extension.dart';
 export 'day_period_extension.dart';
 export 'time_of_day_extension.dart';